Uploaded Automated Invoice system VBA code
This is a cls file which  automates the feature of invoice system for any vendor. The features include like  add a new customer, saves the invoice file in xls and pdf formats. 

To automate the system you have to open your VBA editor in MS-EXCEL and paste this code to enable the above features mentioned.